WebCrawling/Source/requirements.txt
requests
feedparser
python-dotenv
pymysql
beautifulsoup4